antenna driving distance is derived from calibration values for antenna A
we set the delay per 0.10% of antenna extension based on the length of time it takes for the antenna to extend or retract for the full length based on the default speed (255).

a reduction in this speed value from 255 to unknown will result in the antenna being unable to trigger the limit switches at the top and bottom.